packing knapsack: bug fix, segmentation fault caused by strsep()
Segmentation fault can happen calling strsep()
function because a string literal can't be canged.
If a actual string of argument for strsep() is a literal, i.e.

    char *ptr = "4 14";
    strsep(&ptr, " ");

That is why strcep can cause segmentation fault.
So strdup() have to be needed to get modifiable string value
from a string literal.

Signed-off-by: Taeung Song <treeze.taeung@gmail.com>