cheese: Introduce core logic for cheese problem
check_board() decide whether cheese cell(that has cheese)
will be melted or not for one hour.

First of all when scanning the cheese board,
if a cell has cheese check whether contiguous cells
(top, bottom, left, right) of the cell has air.

And then check whether the air cell is ventilated
becase if the air cell is confined with cheese cells,
the air cell is airtight.
(If the air cell is airtight, it can't melt adjoined cheese cells.)

So, in sum, status of cheese cells that are adjoined with air cells
can be changed as 'MELTING_CHEESE' from 'HAS_CHEESE' by check_cheese_cell()

Signed-off-by: Taeung Song <treeze.taeung@gmail.com>