histogram: Infrastructure code to get max area
This program is to get max area of quadrangle in histogram.
So add new four functions as below.

First of all, gather groups per a count that can be from 1 to n
(n mean the number of blocks) and then find max area of each group.
If repeat this process, we can get the final result.

    get_max_area()
	- make_groups()
	- get_max_area_from_group()
	- get_groups_maximum()

Signed-off-by: Taeung Song <treeze.taeung@gmail.com>